- Fixed bug: non-eZ database tables were deleted when selecting remove data in the setup wizard.   http://ez.no/community/bug_reports/other_database_tables_deleted_when_selecting_remove_data_in_setup_wizard
git-svn-id: file:///home/patrick.allaert/svn-git/ezp-repo/ezpublish/stable/3.3@8039 a01eee8c-daba-0310-acae-fa49f3023285

+++ lib/ezdb/classes/ezdbtool.php
@@ -82,6 +82,10 @@ class eZDBTool
             $relationItems = $db->relationList( $relationType );
             foreach ( $relationItems as $relationItem )
             {
+                // skip non-ez relations
+                if ( !eregi( "^ez", $relationItem ) )
+                    continue;
+
                 if ( !$db->removeRelation( $relationItem, $relationType ) )
                 {
                     $result = false;
